Removed unused functions and controllers
