+++ app/Models/Village.php
@@ -4,7 +4,6 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\Facades\DB;
 use App\Helpers\RewardHandler;
 use App\Helpers\LevelHandler;
 use App\Http\Resources\VillageResource;
@@ -36,7 +35,6 @@ class Village extends Model
     }
     public function experienceTable(){
         return VariableHandler::getExperienceTable();
-        return DB::table('experience_points')->get();
     }
 
     /**
