+++ routes/api.php
@@ -7,14 +7,12 @@ use App\Http\Controllers\RegisteredUserController;
 use App\Http\Controllers\TaskListController;
 use App\Http\Controllers\TaskController;
 use App\Http\Controllers\AchievementController;
-use App\Http\Controllers\CharacterController;
 use App\Http\Controllers\NotificationController;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\FriendController;
 use App\Http\Controllers\ExampleTaskController;
 use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\OverviewController;
-use App\Http\Controllers\VillageController;
 use App\Http\Controllers\RewardController;
 
 /*
@@ -49,18 +47,11 @@ Route::group(['middleware' => ['auth']], function () {
     ]);
     Route::post('/tasks/merge/{tasklist}', [TaskListController::class, 'mergeTasks']);
 
-    Route::resource('/character', CharacterController::class)->only([
-        'store', 'update', 'destroy',
-    ]);
+    Route::get('/character/all', [RewardController::class, 'fetchAllCharactersByUser']);
 
-    Route::get('/character/find', [CharacterController::class, 'fetchCharacterIfExists']);
-    Route::get('/character/active', [CharacterController::class, 'fetchActiveCharacter']);
-    Route::get('/character/all', [CharacterController::class, 'fetchAllCharactersByUser']);
-
-    Route::get('/village/all', [VillageController::class, 'fetchAllVillagesByUser']);
+    Route::get('/village/all', [RewardController::class, 'fetchAllVillagesByUser']);
 
     Route::put('/reward/update', [RewardController::class, 'updateRewardObj']);
-
     
     Route::get('/notifications', [NotificationController::class, 'show']);
     Route::get('/notifications/unread', [NotificationController::class, 'hasUnreadNotifications']);
@@ -88,13 +79,13 @@ Route::group(['middleware' => ['auth']], function () {
     Route::get('/overview', [OverviewController::class, 'getOverview']);
 });
 
-    Route::get('/achievements', [AchievementController::class, 'showAll']);
-    Route::get('/achievements/triggers', [AchievementController::class, 'showTriggers']);
-    Route::resource('/achievements', AchievementController::class)->only([
-        'store', 'update',
-    ]);
+Route::get('/achievements', [AchievementController::class, 'showAll']);
+Route::get('/achievements/triggers', [AchievementController::class, 'showTriggers']);
+Route::resource('/achievements', AchievementController::class)->only([
+    'store', 'update',
+]);
 
-    Route::get('/examples/tasks', [ExampleTaskController::class, 'fetchExampleTasks']);
+Route::get('/examples/tasks', [ExampleTaskController::class, 'fetchExampleTasks']);
 //Route::group(['middleware' => ['admin']], function () {
 
 //});
